build(structure): restructure repo with src layout
- introduce pyproject.toml, update tests and all import statements
- fix import paths from lib to src
- add NetCDF test to verify file can be loaded and use pythonpath package
- use xarray[io] instead of xarray[complete]
- bump Pytest version so that pythonpath is compatible
- various src cleanup
