--- requirements.txt
@@ -1,11 +1,0 @@
-numpy==2.0.2
-pandas==2.2.2
-python-dateutil==2.9.0.post0
-pytz==2024.1
-six==1.16.0
-matplotlib==3.9.2
-scipy==1.13.1
-xarray[complete]==2024.7.0
-scikit-learn==1.5.1
-skforecast==0.14.0
-pytest>=8.0.0
