+++ src/nemo_spinup_forecast/forecast_method.py
